server/requirements.txt
aiohttp==3.8.6
aiohttp-index==0.1
python-engineio==3.8.1
python-socketio==4.1.0
scipy==1.14

aiofiles==23.2.1